Inline Disallow paint-over into KeyOptionToggles, remove dead ZonePicker/DisallowPaintOver block
BrushControls is only ever rendered from InspectorControls, which only
mounts when activeTool === INSPECTOR — showZonePicker (gated on
BRUSH/SHATTER/COI-ERASER) was therefore always false there, so that
whole branch (ZonePicker + DisallowPaintOver) never actually rendered.

DisallowPaintOver's remaining live usage was KeyOptionToggles-
exclusive once that dead branch is gone, so it's inlined there now
(preserving its COI-mode hide and its own HelpTip) rather than kept as
a separate one-call component.
